+++ tencentcloud/billing/v20180709/models.go
@@ -987,7 +987,7 @@ type DescribeBillDetailRequestParams struct {
 	// 查询指定资源信息
 	ResourceId *string `json:"ResourceId,omitempty" name:"ResourceId"`
 
-	// 查询交易类型，如下：
+	// 查询交易类型（请使用交易类型名称入参），入参示例枚举如下：
 	// 包年包月新购
 	// 包年包月续费
 	// 包年包月配置变更
@@ -1058,7 +1058,7 @@ type DescribeBillDetailRequest struct {
 	// 查询指定资源信息
 	ResourceId *string `json:"ResourceId,omitempty" name:"ResourceId"`
 
-	// 查询交易类型，如下：
+	// 查询交易类型（请使用交易类型名称入参），入参示例枚举如下：
 	// 包年包月新购
 	// 包年包月续费
 	// 包年包月配置变更