+++ tencentcloud/ckafka/v20190819/models.go
@@ -3599,6 +3599,20 @@ func (r *DescribeCkafkaZoneResponse) FromJsonString(s string) error {
 	return json.Unmarshal([]byte(s), &r)
 }
 
+type DescribeConnectInfoResultDTO struct {
+	// ip地址
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	IpAddr *string `json:"IpAddr,omitempty" name:"IpAddr"`
+
+	// 连结时间
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	Time *string `json:"Time,omitempty" name:"Time"`
+
+	// 是否支持的版本
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	IsUnSupportVersion *bool `json:"IsUnSupportVersion,omitempty" name:"IsUnSupportVersion"`
+}
+
 type DescribeConnectResource struct {
 	// 连接源的Id
 	// 注意：此字段可能返回 null，表示取不到有效值。
@@ -5087,6 +5101,70 @@ func (r *DescribeTopicDetailResponse) ToJsonString() string {
 // FromJsonString It is highly **NOT** recommended to use this function
 // because it has no param check, nor strict type check
 func (r *DescribeTopicDetailResponse) FromJsonString(s string) error {
+	return json.Unmarshal([]byte(s), &r)
+}
+
+// Predefined struct for user
+type DescribeTopicProduceConnectionRequestParams struct {
+	// 实例id
+	InstanceId *string `json:"InstanceId,omitempty" name:"InstanceId"`
+
+	// topic名称
+	TopicName *string `json:"TopicName,omitempty" name:"TopicName"`
+}
+
+type DescribeTopicProduceConnectionRequest struct {
+	*tchttp.BaseRequest
+	
+	// 实例id
+	InstanceId *string `json:"InstanceId,omitempty" name:"InstanceId"`
+
+	// topic名称
+	TopicName *string `json:"TopicName,omitempty" name:"TopicName"`
+}
+
+func (r *DescribeTopicProduceConnectionRequest) ToJsonString() string {
+    b, _ := json.Marshal(r)
+    return string(b)
+}
+
+// FromJsonString It is highly **NOT** recommended to use this function
+// because it has no param check, nor strict type check
+func (r *DescribeTopicProduceConnectionRequest) FromJsonString(s string) error {
+	f := make(map[string]interface{})
+	if err := json.Unmarshal([]byte(s), &f); err != nil {
+		return err
+	}
+	delete(f, "InstanceId")
+	delete(f, "TopicName")
+	if len(f) > 0 {
+		return tcerr.NewTencentCloudSDKError("ClientError.BuildRequestError", "DescribeTopicProduceConnectionRequest has unknown keys!", "")
+	}
+	return json.Unmarshal([]byte(s), &r)
+}
+
+// Predefined struct for user
+type DescribeTopicProduceConnectionResponseParams struct {
+	// 链接信息返回结果集
+	Result []*DescribeConnectInfoResultDTO `json:"Result,omitempty" name:"Result"`
+
+	// 唯一请求 ID，每次请求都会返回。定位问题时需要提供该次请求的 RequestId。
+	RequestId *string `json:"RequestId,omitempty" name:"RequestId"`
+}
+
+type DescribeTopicProduceConnectionResponse struct {
+	*tchttp.BaseResponse
+	Response *DescribeTopicProduceConnectionResponseParams `json:"Response"`
+}
+
+func (r *DescribeTopicProduceConnectionResponse) ToJsonString() string {
+    b, _ := json.Marshal(r)
+    return string(b)
+}
+
+// FromJsonString It is highly **NOT** recommended to use this function
+// because it has no param check, nor strict type check
+func (r *DescribeTopicProduceConnectionResponse) FromJsonString(s string) error {
 	return json.Unmarshal([]byte(s), &r)
 }